Rename Monitor Detours to Client
