+++ Source/Monitor/Client/Functions/LibLoaderApiOverrides.h
