+++ Source/Monitor/Client/Functions/UndocumentedApiCache.h
