+++ Source/Monitor/Client/Functions/UndocumentedApiOverrides.h
