+++ Source/Monitor/Client/Functions/WinBaseCache.h
