+++ Source/Monitor/Client/Functions/WinBaseOverrides.h
